科目知识点开发

+++ src/main/java/com/example/simpleDemo/mapper/SubjectMapper.java
@@ -18,5 +18,5 @@ public interface SubjectMapper {
     int updateSubject(Subject subject);
 
     // 删除科目
-    int deleteSubjectById(@Param("id") Integer id);
+    int deleteSubjectById(@Param("id") Long id);
 }
